+++ packages/proxy/src/routes/address.ts
@@ -17,23 +17,6 @@ export const addressParamsSchema = {
   }),
 };
 
-export type Paginated<T> = {
-  items: T[];
-  meta: {
-    totalItems: number;
-    itemCount: number;
-    itemsPerPage: number;
-    totalPages: number;
-    currentPage: number;
-  };
-  links: {
-    first: string;
-    previous: string;
-    next: string;
-    last: string;
-  };
-};
-
 const transfersSchema = {
   params: z.object({
     address: addressSchema,
